+++ EssentialFeed/EssentialFeed/Feed Cache/Infrastructure/CoreData/ManagedCache.swift
@@ -1,0 +1,30 @@
+//
+//  ManagedCache.swift
+//  EssentialFeed
+//
+//  Created by panditpakhurde on 28/05/20.
+//  Copyright © 2020 Quikr. All rights reserved.
+//
+
+import CoreData
+
+@objc(ManagedCache)
+internal class ManagedCache: NSManagedObject {
+    @NSManaged internal var timestamp: Date
+    @NSManaged internal var feed: NSOrderedSet
+    
+    internal static func newUniqueInstance(in context: NSManagedObjectContext) throws -> ManagedCache {
+        try find(in: context).map(context.delete)
+        return ManagedCache(context: context)
+    }
+    
+    internal static func find(in context: NSManagedObjectContext) throws -> ManagedCache? {
+        let request = NSFetchRequest<ManagedCache>(entityName: entity().name!)
+        request.returnsObjectsAsFaults = false
+        return try context.fetch(request).first
+    }
+    
+    internal var localFeed: [LocalFeedImage] {
+        return feed.compactMap {($0 as? ManagedFeedImage)?.local}
+    }
+}
