+++ EssentialFeed/EssentialFeed/Feed Cache/Infrastructure/CoreData/ManagedFeedImage.swift
@@ -1,0 +1,34 @@
+//
+//  ManagedFeedImage.swift
+//  EssentialFeed
+//
+//  Created by panditpakhurde on 28/05/20.
+//  Copyright © 2020 Quikr. All rights reserved.
+//
+
+import CoreData
+
+@objc(ManagedFeedImage)
+internal class ManagedFeedImage: NSManagedObject {
+    @NSManaged internal var id: UUID
+    @NSManaged internal var imageDescription: String?
+    @NSManaged internal var location: String?
+    @NSManaged internal var url: URL
+    @NSManaged internal var cache: ManagedCache
+    
+    
+    internal static func images(from feed: [LocalFeedImage], in context: NSManagedObjectContext) -> NSOrderedSet {
+        return NSOrderedSet(array: feed.map { local in
+            let managedFeedImage = ManagedFeedImage(context: context)
+            managedFeedImage.id = local.id
+            managedFeedImage.imageDescription = local.description
+            managedFeedImage.location = local.location
+            managedFeedImage.url = local.url
+            return managedFeedImage
+        })
+    }
+    
+    internal var local: LocalFeedImage {
+        return LocalFeedImage(id: id, description: imageDescription, location: location, url: url)
+    }
+}
